backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
mysql-connector-python
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
requests
email-validator
python-dotenv
gunicorn
